Move project files to src/internal
