+++ packages/hardhat-verify/src/internal/solc/artifacts.ts
