+++ packages/hardhat-verify/src/internal/solc/metadata.ts
